refactor(mcp): 统一 ToolCallResult 类型定义，明确 MCP 管理器分层
- 将 ToolCallResult canonical 定义收敛到 mcp-core/types.ts，消除三处重复定义
- SDK 原始类型改名为 CoreToolCallResult，自定义 interface 作为主要导出
- server/lib/mcp/types.ts 和 endpoint/types.ts 改为从 mcp-core re-export
- 删除 mcp-core 中混淆的 MCPServiceManager 别名导出
- 更新 MCPManager 类注释，明确定位为协议层管理器（轻量级）
- 修复 content 可选后的类型收窄问题
